refactor : ♻️ 구매할 수 있는 상품 현황 테이블 render하는 함수 분리 기존에 renderProductPurchaseMenu에서 renderNowSellingProductTable.js에서 값을 받아와 렌더링하던 방식에서 renderNowSellingProductTable에서 구매할 수 있는 상품 현황 요소들을 만들고 렌더링까지 하도록 변경  renderNowSellingProductTable.js / renderProductPurchaseMenu.js
